Change ints to lls where it makes sense

+++ src/cpp_src/pb_methods_and_measures/Greedy.cpp
@@ -10,7 +10,7 @@
 #include <vector>
 
 std::vector<ProjectEmbedding> greedy(const Election &election, const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
+    long long total_budget = election.budget();
     auto projects = election.projects();
     std::vector<ProjectEmbedding> winners;
     std::ranges::sort(projects, [&tie_breaking](ProjectEmbedding a, ProjectEmbedding b) {
@@ -30,12 +30,13 @@ std::vector<ProjectEmbedding> greedy(const Election &election, const ProjectComp
     return winners;
 }
 
-std::optional<int> cost_reduction_for_greedy(const Election &election, int p, const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
+std::optional<long long> cost_reduction_for_greedy(const Election &election, int p,
+                                                   const ProjectComparator &tie_breaking) {
+    long long total_budget = election.budget();
     auto projects = election.projects();
     auto pp = projects[p];
 
-    std::optional<int> max_price_to_be_chosen{};
+    std::optional<long long> max_price_to_be_chosen{};
 
     std::ranges::sort(projects, [&tie_breaking](ProjectEmbedding a, ProjectEmbedding b) {
         if (a.approvers().size() == b.approvers().size()) {
@@ -53,7 +54,7 @@ std::optional<int> cost_reduction_for_greedy(const Election &election, int p, co
                 return pp.cost();
             }
             if (project.approvers().size() == pp.approvers().size()) { // Not taken because lost tie-breaking
-                int current_max_price = 0;
+                long long current_max_price = 0;
                 if (tie_breaking(ProjectEmbedding(project.cost(), pp.name(), pp.approvers()), project)) {
                     current_max_price = std::max(current_max_price, project.cost());
                 }
@@ -70,9 +71,10 @@ std::optional<int> cost_reduction_for_greedy(const Election &election, int p, co
     return max_price_to_be_chosen;
 }
 
-std::optional<int> optimist_add_for_greedy(const Election &election, int p, const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
-    int num_voters = election.numVoters();
+std::optional<long long> optimist_add_for_greedy(const Election &election, int p,
+                                                 const ProjectComparator &tie_breaking) {
+    long long total_budget = election.budget();
+    long long num_voters = election.numVoters();
     auto projects = election.projects();
     auto pp = projects[p];
     if (pp.cost() > total_budget)
@@ -90,7 +92,7 @@ std::optional<int> optimist_add_for_greedy(const Election &election, int p, cons
                 return 0;
             }
             if (pp.cost() > total_budget - project.cost()) { // if (last moment to add pp)
-                int new_approvers_size = project.approvers().size();
+                long long new_approvers_size = project.approvers().size();
                 std::vector<int> new_approvers(new_approvers_size);
                 std::iota(new_approvers.begin(), new_approvers.end(), 0);
                 auto new_pp = ProjectEmbedding(pp.cost(), pp.name(), new_approvers);
@@ -108,12 +110,14 @@ std::optional<int> optimist_add_for_greedy(const Election &election, int p, cons
     return {}; // LCOV_EXCL_LINE (every project should be feasible)
 }
 
-std::optional<int> pessimist_add_for_greedy(const Election &election, int p, const ProjectComparator &tie_breaking) {
+std::optional<long long> pessimist_add_for_greedy(const Election &election, int p,
+                                                  const ProjectComparator &tie_breaking) {
     return optimist_add_for_greedy(election, p, tie_breaking);
 }
 
-std::optional<int> singleton_add_for_greedy(const Election &election, int p, const ProjectComparator &tie_breaking) {
-    int total_budget = election.budget();
+std::optional<long long> singleton_add_for_greedy(const Election &election, int p,
+                                                  const ProjectComparator &tie_breaking) {
+    long long total_budget = election.budget();
     auto projects = election.projects();
     auto pp = projects[p];
     if (pp.cost() > total_budget)
@@ -131,7 +135,7 @@ std::optional<int> singleton_add_for_greedy(const Election &election, int p, con
                 return 0;
             }
             if (pp.cost() > total_budget - project.cost()) { // if (last moment to add pp)
-                int new_approvers_size = project.approvers().size();
+                long long new_approvers_size = project.approvers().size();
                 std::vector<int> new_approvers(new_approvers_size);
                 std::iota(new_approvers.begin(), new_approvers.end(), 0);
                 auto new_pp = ProjectEmbedding(pp.cost(), pp.name(), new_approvers);
