+++ src/cpp_src/pb_methods_and_measures/GreedyOverCost.h
@@ -7,14 +7,14 @@
 
 std::vector<ProjectEmbedding> greedy_over_cost(const Election &election, const ProjectComparator &tie_breaking);
 
-std::optional<int> cost_reduction_for_greedy_over_cost(const Election &election, int p,
-                                                       const ProjectComparator &tie_breaking);
+std::optional<long long> cost_reduction_for_greedy_over_cost(const Election &election, int p,
+                                                             const ProjectComparator &tie_breaking);
 
-std::optional<int> optimist_add_for_greedy_over_cost(const Election &election, int p,
-                                                     const ProjectComparator &tie_breaking);
+std::optional<long long> optimist_add_for_greedy_over_cost(const Election &election, int p,
+                                                           const ProjectComparator &tie_breaking);
 
-std::optional<int> pessimist_add_for_greedy_over_cost(const Election &election, int p,
-                                                      const ProjectComparator &tie_breaking);
+std::optional<long long> pessimist_add_for_greedy_over_cost(const Election &election, int p,
+                                                            const ProjectComparator &tie_breaking);
 
-std::optional<int> singleton_add_for_greedy_over_cost(const Election &election, int p,
-                                                      const ProjectComparator &tie_breaking);
+std::optional<long long> singleton_add_for_greedy_over_cost(const Election &election, int p,
+                                                            const ProjectComparator &tie_breaking);
