+++ src/cpp_src/pb_methods_and_measures/MesApr.h
@@ -7,4 +7,5 @@
 
 std::vector<ProjectEmbedding> mes_apr(const Election &election, const ProjectComparator &tie_breaking);
 
-std::optional<int> cost_reduction_for_mes_apr(const Election &election, int p, const ProjectComparator &tie_breaking);
+std::optional<long long> cost_reduction_for_mes_apr(const Election &election, int p,
+                                                    const ProjectComparator &tie_breaking);
