+++ src/cpp_src/utils/ProjectEmbedding.h
@@ -9,21 +9,21 @@ class ProjectComparator;
 class ProjectEmbedding {
   public:
     template <typename StringT, typename VectorT>
-    ProjectEmbedding(int cost, StringT &&name, VectorT &&approvers)
+    ProjectEmbedding(long long cost, StringT &&name, VectorT &&approvers)
         : cost_(cost), name_(std::forward<StringT>(name)), approvers_(std::forward<VectorT>(approvers)) {}
 
     template <typename StringT>
-    ProjectEmbedding(int cost, StringT &&name) : cost_(cost), name_(std::forward<StringT>(name)) {}
+    ProjectEmbedding(long long cost, StringT &&name) : cost_(cost), name_(std::forward<StringT>(name)) {}
 
     bool operator==(const ProjectEmbedding &other) const { return name_ == other.name_; }
-    int cost() const { return cost_; }
+    long long cost() const { return cost_; }
     const std::string &name() const { return name_; }
     const std::vector<int> &approvers() const { return approvers_; }
 
     friend class ProjectComparator;
 
   private:
-    int cost_;
+    long long cost_;
     std::string name_;
     std::vector<int> approvers_;
 };
