+++ src/main.cpp
@@ -29,7 +29,7 @@ PYBIND11_MODULE(_core, m) {
         .finalize();
 
     py::class_<ProjectEmbedding>(m, "ProjectEmbedding")
-        .def(py::init<int, std::string, std::vector<int>>(), "cost"_a, "name"_a, "approvers"_a)
+        .def(py::init<long long, std::string, std::vector<int>>(), "cost"_a, "name"_a, "approvers"_a)
         .def(py::init<int, std::string>(), "cost"_a, "name"_a)
         .def_property_readonly("cost", &ProjectEmbedding::cost)
         .def_property_readonly("name", &ProjectEmbedding::name)
@@ -52,7 +52,7 @@ PYBIND11_MODULE(_core, m) {
                                       [](py::object) { return ProjectComparator::ByCostDescThenVotesDesc; });
 
     py::class_<Election>(m, "Election")
-        .def(py::init<int, int, std::vector<ProjectEmbedding>>(), "budget"_a, "num_voters"_a, "projects"_a)
+        .def(py::init<long long, long long, std::vector<ProjectEmbedding>>(), "budget"_a, "num_voters"_a, "projects"_a)
         .def_property_readonly("budget", &Election::budget)
         .def_property_readonly("num_voters", &Election::numVoters)
         .def_property_readonly("projects", &Election::projects);
